Task 4: Refactoring
